backend/requirements.txt
flask
flask-cors
minio
python-dotenv
